[GILJOB-12] 커스텀 에러 코드 및 ErrorResponse 분리
