+++ src/main/kotlin/com/yapp/giljob/global/error/ErrorCode.kt
@@ -1,0 +1,15 @@
+package com.yapp.giljob.global.error
+
+import org.springframework.http.HttpStatus
+
+enum class ErrorCode(
+    val status: HttpStatus,
+    val code: String,
+    val message: String
+) {
+    // Common
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "C001", "Internal Server Error"),
+    INVALID_INPUT_VALUE(HttpStatus.BAD_REQUEST, "C002", "Invalid Input Value"),
+    METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, "C003", "Method Not Allowed"),
+    ENTITY_NOT_FOUND(HttpStatus.NOT_FOUND, "C004", " Entity Not Found");
+}
