+++ src/main/kotlin/com/yapp/giljob/global/error/exception/BusinessException.kt
@@ -1,0 +1,7 @@
+package com.yapp.giljob.global.error.exception
+
+import com.yapp.giljob.global.error.ErrorCode
+
+open class BusinessException(
+    val errorCode: ErrorCode
+) : RuntimeException()
