--- src/main/kotlin/com/yapp/giljob/global/exception/BusinessException.kt
@@ -1,5 +1,0 @@
-package com.yapp.giljob.global.exception
-
-open class BusinessException(
-    val exceptionCode: ExceptionCode
-) : RuntimeException()
