--- src/main/kotlin/com/yapp/giljob/global/exception/ExceptionCode.kt
@@ -1,10 +1,0 @@
-package com.yapp.giljob.global.exception
-
-import org.springframework.http.HttpStatus
-
-enum class ExceptionCode(
-    val status: HttpStatus,
-    val message: String
-) {
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "Internal Server Error");
-}
